+++ R/bpipe.R
@@ -1,4 +1,65 @@
 
+
+#' Pipe step operator
+#'
+#' @param pipe_left_arg left argument
+#' @param pipe_right_arg substitute(pipe_right_arg) argument
+#' @param pipe_environment environment to evaluate in
+#' @return result
+#'
+#' @export
+#'
+pipe_step <- function(pipe_left_arg, pipe_right_arg,
+                      pipe_environment) {
+  UseMethod("pipe_step", pipe_left_arg)
+}
+
+#' Pipe step operator
+#'
+#' @param pipe_left_arg left argument
+#' @param pipe_right_arg substitute(pipe_right_arg) argument
+#' @param pipe_environment environment to evaluate in
+#' @return result
+#'
+#' @export
+#'
+pipe_step.default <- function(pipe_left_arg, pipe_right_arg,
+                              pipe_environment) {
+  # eval by with pipe_left_arg's value in dot (simulates chaining)
+  assign(".", pipe_left_arg,
+         envir = pipe_environment,
+         inherits = FALSE)
+  eval(pipe_right_arg,
+       envir = pipe_environment,
+       enclos = pipe_environment)
+}
+
+#' Wrapr function.
+#'
+#' S3 dispatch on tyhpe of pipe_right_argument.
+#'
+#' @param pipe_left_arg left argument
+#' @param pipe_right_arg right argument
+#' @param pipe_environment environment to evaluate in
+#' @return result
+#'
+#' @export
+#'
+wrapr_function <- function(pipe_left_arg, pipe_right_arg,
+                      pipe_environment) {
+  UseMethod("wrapr_function", pipe_right_arg)
+}
+
+
+#' Pipe implementation.
+#'
+#' @param pipe_left_arg substitute(pipe_left_arg) argument
+#' @param pipe_right_arg substitute(pipe_right_arg) argument
+#' @param pipe_environment environment to evaluate in
+#' @return result
+#'
+#' @noRd
+#'
 pipe_impl <- function(pipe_left_arg, pipe_right_arg, pipe_environment) {
   # force pipe_left_arg, by left-associativity "pipe_left_arg" may be a pipe
   # sequence itself.
@@ -7,10 +68,6 @@ pipe_impl <- function(pipe_left_arg, pipe_right_arg, pipe_environment) {
   pipe_left_arg <- eval(pipe_left_arg,
                         envir = pipe_environment,
                         enclos = pipe_environment)
-  # eval by with pipe_left_arg's value in dot (simulates chaining)
-  assign(".", pipe_left_arg,
-         envir= pipe_environment,
-         inherits= FALSE)
   # special case: dereference names
   if(is.name(pipe_right_arg)) {
     v <- base::mget(as.character(pipe_right_arg),
@@ -21,7 +78,7 @@ pipe_impl <- function(pipe_left_arg, pipe_right_arg, pipe_environment) {
       pipe_right_arg <- v
     }
   }
-  # special case: functions
+  # special case: functions (S3 doesn't do this)
   if(is.function(pipe_right_arg)) {
     return(do.call(pipe_right_arg,
                    list(pipe_left_arg),
@@ -30,18 +87,13 @@ pipe_impl <- function(pipe_left_arg, pipe_right_arg, pipe_environment) {
   # special case: look for wrapr_applicable objects
   if((!is.atomic(pipe_right_arg)) &&
      ("wrapr_applicable" %in% class(pipe_right_arg))) {
-    f <- pipe_right_arg$wrapr_function
-    if((!is.null(f)) && (is.function(f))) {
-      return(do.call(f,
-                     list(pipe_left_arg = pipe_left_arg,
-                          pipe_right_arg = pipe_right_arg,
-                          pipe_environment = pipe_environment),
-                     envir = pipe_environment))
-    }
+    # S3 dispatch on right argument
+    return(wrapr_function(pipe_left_arg,
+                          pipe_right_arg,
+                          pipe_environment))
   }
-  eval(pipe_right_arg,
-       envir = pipe_environment,
-       enclos = pipe_environment)
+  # Go for S3 dispatch
+  pipe_step(pipe_left_arg, pipe_right_arg, pipe_environment)
 }
 
 #' Pipe operator ("dot arrow").